refactor media-upload-form into media-upload-modal
