--- src/app/assets/scss/components/_media-upload-form.scss
@@ -1,8 +1,0 @@
-// src/app/views/river-detail/shared/components/media-upload-form.vue
-
-.media-upload-form {
-
-  .bx--modal .bx--text-input {
-    background-color: #f4f7fb;
-  }
-}
