+++ src/app/assets/scss/components/components.scss
@@ -2,8 +2,6 @@
 
 @import "./river-table";
 
-@import "./media-upload-form";
-
 @import "./image-gallery";
 
 @import "./utility-block";
