+++ src/app/views/river-detail/components/index.js
@@ -1,7 +1,6 @@
 export { default as AlertUpdateModal } from './alert-update-modal.vue'
 export { default as ImageGallery } from './image-gallery/image-gallery.vue'
 export { default as MapBanner } from './map-banner.vue'
-export { default as MediaUploadForm } from './media-upload-form.vue'
 export { default as MediaUploadModal } from './media-upload-modal.vue'
 export { default as ReachTitleEditModal } from './reach-title-edit-modal.vue'
